Fixed the query being sent out on load of the Punjab, Pakistan subregion.
The issue was that there were two subregions named Punjab: one in India and one in Pakistan. The results were sorted alphabetically by country name, and the first one was always being used. This commit adds an extra parameter into the query to query for subregions and country names, to ensure a more precise result.

To confirm that we did not break any other regions, we ran a query on the mapservice https://gis-gfw.wri.org/arcgis/rest/services/admin/MapServer/5 for distinct values of NAME_0 and noted no null or empty values.